backend/src/auth/auth.service.ts
import {
  Injectable,
  UnauthorizedException,
  UnprocessableEntityException,
} from '@nestjs/common';
import { JwtService } from '@nestjs/jwt';
import { UserService } from '../user/user.service';
import { User } from '../user/user.entity';
import { AuthUserResponse } from './responses/auth-user.response';
import { RegisterUserInput } from './inputs/register-user.input';

@Injectable()
export class AuthService {
  constructor(
    private readonly jwtService: JwtService,
    private readonly userService: UserService,
  ) {}

  async validateCredentials(username: string, password: string): Promise<User> {
    const user = await this.userService.findOneByUsername(username);
    if (!(await user?.comparePassword(password))) {
      throw new UnauthorizedException();
    }
    return user;
  }

  async signToken(user: User): Promise<AuthUserResponse> {
    const payload = { username: user.username, sub: user.id };
    return new AuthUserResponse({
      user,
      token: this.jwtService.sign(payload),
    });
  }

  async registerUser(user: RegisterUserInput): Promise<AuthUserResponse> {
    if (await this.userService.existsByCredentials(user)) {
      throw new UnprocessableEntityException();
    }

    const returnedUser = await this.userService.save(user);
    return this.signToken(returnedUser);
  }
}
